requirements.txt
pandas
psycopg2-binary
PyYAML
scikit-learn
torch       # CPU version (safe on all machines)
tqdm